82-86.Routing
There were discussed react-router-dom library, use of index atribute in Route, what Outlet is used for, how to nest Route components
