+++ crwn-clothing/package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "6",
     "react-scripts": "5.0.1",
     "sass": "^1.85.1",
     "web-vitals": "^2.1.4"
